Feature : item/pause menu
Known Issue : When key press is short, it is sometimes not registered and the menu doesn't open/close

When user inputs the proper key (ESCAPE at the moment), he shows/hides the item/pause menu.

This menu displays the items owned (for test purposes, right now every item that I intend on developing is flagged as owned, but they are not implemented yet), both the passive and the active ones.

The equipped main weapon (currently the wood sword) is now visible in the HUD in both level view and item/pause menu, in the A slot.

Item selection for slot B is not part of the scope of this feature

The triforce on the bottom half of the item/pause menu isn't implemented, the "empty" version of it is currently part of the background of the menu