Feature - Loot Consumables
Hearts, Bombs, Rupees and Fairies are now actually dropped when a monster drops an item, corresponding to the loot table.

A Heart heals one full heart of the player's health, up to the current max
A Rupee adds one rupee to the player's cash stash, up to 999
A Bombs adds 3 bombs to the player's bomb reserve, up to max (15 configurable)
A Fairy heals the player to the PLAYER_HEALTH_MAX, but you have to catch it as it flies around randomly

Yes, I know, there is a spelling mistake in the branch name. Shame.