Feature - Item Selector
When in item menu, you now can use left and right arrows (or q/d) to choose which item to use with Action B (left Shift)
Chosen item is set when leaving the menu.

Player starts with 4 bombs to be able to test the feature.
When the player doesn't have any item at all, the selector hovers the first item position (currently Boomerang). There is no up/down motion if new items are added, bottom line is reached using RIGHT at the end of the top line.
Selector doesn't loop around, LEFT on the first item or RIGHT on the last doesn't move it.
Selector jumps from owned item to owned item, player is not supposed to be able to set no item as item B.

Player will start with no item later in the dev (when there will be items other than Bombs to loot, such as the Candle or the Boomerang)