Refacto - Player Death Animation:
Some death animations were handled in Level despite animating the Player, which was utterly disgusting. Now it's done in the Player object through proper calls and state handling.

Hurt animation when dying is a separate animation that loops forever (to separate it from hurt state, where Player moves at a speed relative to the hurt frame, since if I add the ability to loop on it, player is pushed back strongly at the very end of the hurt animation)

Spin animation is now part of the Player class, handled as a proper state.

"red" floors now have a configurable name available in settings.py, and the par of the code shifting between these backgrounds is now much cleaner and less copy/pasted