Feature - level scrolling
Player can use invisible warps on the open sides of the screen he's navigating on.
This causes a scrolling animation to show the new screen.
Player can also take warps on specific spots of the map to enter caves / dungeons.
Caves configured as such will have a Stairs animation & sound
Maps accessed through (future) stairs sprites on the ground (such as secret under trees) will make no sound nor animation, as it is done this way in the NES game