Feature - Ladder
Player can use its ladder (if owned). The ladder allows the player to walk on a tile of water, by putting it above said water tile.
This allows the crossing of small bodies of water (32*32px, 2 water tiles large, same size as the player)
Ladder is removed when the player is not on top of it anymore.