Feature - item pick up
A Dictionary of dictionaries MAP_ITEMS has been defined in settings.py to describe item content of each level that contains any.
load_items handles the proper checks to ensure that if an item id is defined in the csv, but not in the MAP_ITEMS, game doesn't crash, the item is ignored and not placed in the level screen

Items of this sort now take the level_id as a parameter to their constructor, in order to change the MAP_ITEM when pick up through .effect(). This way, an item that the Player picked up will not be spawned again when the map screen is loaded again