Refacto - debug
Removed references to the debug function that displays text & info on the screen for dev uses.

The removed instances were for cases which should not happen unless someone implements new behaviours, and shouldn't be displayed as text.

Replaced with either nothing if it breaks nothing, or returns from the function if it's an illegal/not yet implemented option

General cleanup of these kind of situations.