Feature - Boomerang
Player can now loot and use the boomerang item (temporarily available in the cave)

On its way forward, can hit enemies, or monster loot. will start going back towards the Player as soon as it hits any of these, or the screen border.
It damages enemies, and loots items for the Player.
It is not bothered by any obstacle (rock, tree, ...) and flies above it.

It's thrown in the same direction the Player is walking to, or if not moving, the same direction he's facing. It follows the Player movements when coming back.

I refactored how Enemy and Particle interact. Now, hurting an Enemy is the Particle's job. Enemy doesn't check Particles anymore in their collisions.

I also refactored the pickup system of items, it's better handled, and now picking up an item switches the equipped item with the new one instantly.