Feature - Candle
Player can use his candle to lit fire.
For now, candle is available in the cave.

Flame spawns from the player's hand, travels forward for a bit, and stays lit for a short while. Then is extinct.

Flame hurts monsters, doesn't affect loot, and will, in time, allow to burn secret trees to discover secret passages.
Player is immune to fire, because why not.