Feature - Destructible Tiles
Bombs can now destroy destructible tiles to reveal hidden passages. For now, caves, that play the stairs animation & sound, then warp you to the new screen.
Flames can now burn destructible tiles to reveal hidden passages. It reveals stairs, that warp you to the new screen without the stairs animation & sound (weird, but ISO to the NES game)

Revealed sprites disappear during side screen scrolling. I managed to make it scroll with the screen when the player LEAVES a screen, but in the case you ENTER a screen with a revealed secret, since the level sprites are loaded only after the scroll animation is played, I decided to not implement this secret scrolling.

secrets revealed are memorized through MAP_SECRETS_REVEALED when revealed, in the form 'level_id': True
This boolean value is checked when loading the secrets_flame or secrets_bomb files.

There is a tree you can burn in the topmost screen, and a cave you can bomb in the rightmost screen.