Feature - Shop
The system for shops has been implemented.
Shops are rooms, that can be kept by a npc, and that offer from 1 up to 3 items on display for the player to pick and pay, with a possibility of displaying text at the top of the room.
Price is shown if the item is not free or negative amount, or if the item is not rupees.
Once an item is bought, it never appears again.
Placement of items depends on the number of items on display

This allows to :
Create rooms that give the player an item for free (sword cave, it's dangerous to go alone)
Create shop rooms that sell items to the player for a shown price
Create rooms that gift the player money (It's a secret to everybody)
Create rooms that collects player's money (PAY FOR MY DOOR !!!)

The text system has been rebuilt entirely, it was more than rudimentary for the "game over" text, now it's a proper class and method, displayed like any other sprite, allows for proper handling of multiple lines, will not cut words in half, ...