Refacto - Player animate()
Refactored player.animate() to delete the repetition of the same code with variations mostly in variable names, and slightly in behaviour (reset or not, and idle after animation or not).

It now has its own change_animation_frame() to compute which is the next frame to paint.

On the side, a bug was introduced that mostly didn't affect anything, EXCEPT the player.kill() when the player death animation was done playing.
As a note : Sprite.kill() only removes the sprite from all its groups. That's all. No destruction in memory, nothing, garbage collection will do it at some point.
So when I implemented the "visible_groups.add(Player)" (that I forgot about) in the level.run() loop to ensure Player sprite is always on top of later created sprites, I completely rendered the player.kill() useless in death, because it was ALWAYS re added to the visible group.
To fix it, I moved the add(Player) from the general run() case to the if condition where the player is still alive, which corrects everything, and now the player doesn't shine like a star on the game over screen.