Fix - Bomb secrets
Bomb secrets were revealed when touched by Bomb directly, not when Bomb is exploding : this is caused by calling super().update() which by default calls for collision('') before we want it.

Bomb explosion hitbox was not centered where it should have been, as rect moved depending on the direction the player was facing, and hitbox was defined before this alteration of position.

Warp hitbox encompasses the whole rect size (2*TILE_SIZE, 2*TILE_SIZE) when not destroyed, and when revealed is shrunk to the proper size in order to have the player animate at the proper position

Explosion hitbox has been modified to better fit visually with the smoke effects and general feel.