Refacto - Constant Str as settings.py global 'const'
There were an unhealthy amount of directions, statuses, paths, labels and others littered everywhere, that would have been hell to modify, or easy to mistype somewhere in any kind of further development or re factoring.

Now, everything Str is defined in settings.py, and used everywhere else. This is commented as good as I could, to complement names that I hope are clear enough