Feature - Blue Octorock
New Enemy : Blue Octorock
Just like the Red Octorock, but with more health, twice as lethal during collision, and with rocks that are generally aimed at the player, not randomly. Aimed on one of the x/y-axis, not both like the Zora, though.
Also, attacks twice as often.

Enemy attack frequency is now a parameter available in settings.py