Feature - Stalfos
Dungeon spooky skeleton. Kind of dumb, but durable, and it has a sword !
Doesn't attack, just damage by collision.
It moves randomly for some time, then remembers the Player here, and go straight toward the Player in a straight line, too dumb to navigate around obstacles.
When Player hits it, it makes it forget about the Player for a few seconds, before it comes back to attack you again.