Feature - Correct dungeon Bomb Secrets
When a hole is made in a wall in the dungeon, the proper "hole in the wall" sprite is now displayed. Only one version of the sprite, the upright one, and it is rotated depending on the side it's supposed to go to.

Warps tile set image has been updated.
There was also a slight error in this level4 png file : a fine gray line was visible in the hole in the right wall. This is now rectified.