Feature - Monster killed event
Maps can implement an event when all Enemies within the screen are killed by defining a MONSTER_KILL_EVENT dict entre { level_id: label }
label can be one of the ITEM_LABEL checked in level.map_kill_event_done(), or it can be OPEN_DOORS_LABEL which is not implemented yet, but later will kill all closed_event_door_sprites.

dungeon0_9 has been modified to accommodate with the fact that I chose to drop the event item right in the middle of the level screen.