Feature - Physicalized player Shield
Player shield collision detection worked for octorock / moblin projectiles that were only in a single direction, X or Y.
Didn't work for Goriya's diagonal boomerangs !

Now, there is a hitbox to represent the shield, Player.shield_hitbox, that is 6px thick, and 4px shorter on each side than the player. If a projectile that should hurt the player, and doesn't ignore shield hits this hitbox, it's parried.