Refacto - Sprite groups
I used to send all groups to the Sprite constructor through n levels of inheritance as lists of Sprite.Group(), which was kind of not the correct way. It worked, but wasn't pretty, and raised a whole lot of syntax warning regarding expected types.

I now send tuples instead of groups, and I do not send them to Sprite.__init__, as it expects * groups as *args, not a list/tuple of groups.
Since all Sprite.__init__ does is to do sprite.add(group), I now loop on the tuple elements to do the same in my constructors that inherit from Sprite.