Feature - Dungeon Boss
The Darknut protects the Triforce ! And it's a tough nut to crack.

Heavy armored, it takes reduced damages ! The flame is thus ineffective against it.
It also has a shield ! Sword and Boomerang won't hurt it if you hit it from the front !
A Bomb explosion seem to shock it into a stun for a few seconds. Then it can't hold its shield, it is vulnerable ! Pain wakes it from its stun though.

It has strong magic : it shoots 3 magic bullets, one directly at you, and two at +/- 45 degrees angles.
It's not fast, but can dash toward you for a few meters ! When dashing, it can't follow you, it runs straight at where you were when it started.

What a beast ! You'd better take a lot of bombs with you.