Refacto - Code as module
main is out of the code directory, and the code folder is a module now, so main can import its components.

On the side, fixed an oversight where revealed bomb secrets sprites wouldn't be killed when loading, and also removed the possibility of stunning the boss during his (albeit very short) hurt animation.