Feature - Executable File
For Windows users, an .exe file has been produced.
It will very likely be flagged as a false positive threat for the user.

This is due to the fact that many malwares are made with PyInstaller, and the only common factor between these malwares is the parts of PyInstaller's bootloader. Thus, anything done with PyInstaller might get flagged as a false positive.
Except if I had a signed certificate to sign my executable file, which won't happen.