Feature - Separate hitbox
Player hitbox has been separated in hitbox and battle_hitbox.

battle_hitbox is used for physical collision with Enemy sprites, for everything else, hitbox is used.

battle_hitbox is a 16x16 pixel square at the center of the Player sprite.

This resolves the imbalance between hitting with the slim and short reaching sword against getting hit as soon as a monster touches you.